(1)Move new s3 code to new module (2)some updated methods
